requirements.txt
#pysqlite3-binary
openai
google-generativeai
python-dotenv
flask
langchain
langchain-community
langchain-google-genai
langchain-chroma
streamlit
requests
pandas
metapub
lxml-html-clean
chromadb
torch
transformers
datasets
sentence-transformers
pandas
numpy
nltk
pymupdf4llm
biopython
Beautifulsoup4
tiktoken
openpyxl
rouge
torchtext
bert_score
sagemaker
boto3
deepeval
refchecker